desvided detail section into reusable components
